python_utils/requirements.txt
#--extra-index-url http://admin:nevarocks@20.0.5.164:8080/simple/ --trusted-host 20.0.5.164
#impyla==0.14.2+1.ge487fd5
pykafka==2.8.0
PyMySQL==0.9.2
redis==2.10.6
redis-py-cluster==1.3.5
slackclient==2.1.0
SQLAlchemy==1.2.14
hvac==0.7.2
PyYAML==5.1
requests==2.21.0
